Eliminated compilation warnings from Thing plugin.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@12156 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/thing/object/curve.cpp
@@ -599,9 +599,9 @@ void csCurve::InitializeDefaultLighting ()
   LightMap->Alloc (
       CURVE_LM_SIZE * csLightMap::lightcell_size,
       CURVE_LM_SIZE * csLightMap::lightcell_size,
-      	ambient.red * 255.0f,
-      	ambient.green * 255.0f,
-      	ambient.blue * 255.0f);
+      	int(ambient.red * 255.0f),
+      	int(ambient.green * 255.0f),
+      	int(ambient.blue * 255.0f));
   LightmapUpToDate = false;
 }
 
@@ -616,9 +616,9 @@ bool csCurve::ReadFromCache (iFile* file)
   LightMap->Alloc (
       CURVE_LM_SIZE * csLightMap::lightcell_size,
       CURVE_LM_SIZE * csLightMap::lightcell_size,
-      	ambient.red * 255.0f,
-      	ambient.green * 255.0f,
-      	ambient.blue * 255.0f);
+      	int(ambient.red * 255.0f),
+	int(ambient.green * 255.0f),
+      	int(ambient.blue * 255.0f));
 
   LightMap->ReadFromCache (
       file,
